requirements.txt
fastapi
uvicorn
python-multipart
apify-client
openai
pydantic
python-dotenv
qdrant-client
requests